- Modified database column - Refactored MailHandler, DBHandler structure - Changed sensor data generation

+++ CS595database.sql
@@ -20,7 +20,7 @@ INSERT INTO DEVICE_MOTION (Device_id, Patient_id, Patient_Name, DO_AX1, DO_AX2, 
 INSERT INTO DEVICE_MOTION (Device_id, Patient_id, Patient_Name, DO_AX1, DO_AX2, DO_AX3, DO_GX1, DO_GX2, DO_GX3, DATE_TIME) VALUES (6, 128, 'Ron', 2,1,2,1,2,3, '2015:07:25 16:45:10');
 INSERT INTO DEVICE_MOTION (Device_id, Patient_id, Patient_Name, DO_AX1, DO_AX2, DO_AX3, DO_GX1, DO_GX2, DO_GX3, DATE_TIME) VALUES (7, 129, 'Eric', 2,2,1,1,2,3, '2015:07:26 18:25:20');
 
-CREATE TABLE FALL_DATA (Fall_id INT (10), Fall_Type varchar  (30), Fall_Time TIMESTAMP, Patient_id INT (10));
+CREATE TABLE FALL_DATA (Fall_id BIGINT, Fall_Type varchar  (30), Fall_Time TIMESTAMP, Patient_id INT (10));
 INSERT INTO FALL_DATA (Fall_id, Fall_Type, Fall_Time, Patient_id) VALUES(131, 'Not Adverse', '2015-07-20 10:30:20',  123);
 INSERT INTO FALL_DATA (Fall_id, Fall_Type, Fall_Time, Patient_id) VALUES(132, 'Not Adverse', '2015-07-21 15:20:30',  124);
 INSERT INTO FALL_DATA (Fall_id, Fall_Type, Fall_Time, Patient_id) VALUES(133, 'Not Adverse', '2015-07-22 04:25:30',  125);
